+++ packages/components/CHANGELOG.md
@@ -1,3 +1,89 @@
+# v20.37.1 (Wed Dec 20 2023)
+
+#### 🐛 Bug Fix
+
+- Bugfix: Error-message overlapping with succeeding element [#813](https://github.com/Infineon/infineon-design-system-stencil/pull/813) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v20.37.0 (Tue Dec 19 2023)
+
+#### 🚀 Enhancement
+
+- Enhancement: Tab to be set from outside, Keyboard navigation [#741](https://github.com/Infineon/infineon-design-system-stencil/pull/741) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v20.36.5 (Mon Dec 18 2023)
+
+#### 🐛 Bug Fix
+
+- 766 multi select dynamic loading + filtering [#775](https://github.com/Infineon/infineon-design-system-stencil/pull/775) (lechneve@ISCN5CG1201S7M.infineon.com [@tishoyanchev](https://github.com/tishoyanchev) [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 3
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v20.36.4 (Mon Dec 18 2023)
+
+#### 🐛 Bug Fix
+
+- updated icons-library dependency [#829](https://github.com/Infineon/infineon-design-system-stencil/pull/829) (lechneve@ISCN5CG1201S7M.infineon.com [@verena-ifx](https://github.com/verena-ifx))
+
+#### Authors: 2
+
+- [@verena-ifx](https://github.com/verena-ifx)
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
+# v20.36.3 (Thu Dec 07 2023)
+
+#### 🐛 Bug Fix
+
+- Icon-button: size inconsistency [#739](https://github.com/Infineon/infineon-design-system-stencil/pull/739) ([@tishoyanchev](https://github.com/tishoyanchev))
+
+#### Authors: 1
+
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+
+---
+
+# v20.36.2 (Thu Dec 07 2023)
+
+#### 🐛 Bug Fix
+
+- Bugfix: disabled check added to button component [#794](https://github.com/Infineon/infineon-design-system-stencil/pull/794) (lechneve@ISCN5CG1201S7M.infineon.com [@tishoyanchev](https://github.com/tishoyanchev) [@verena-ifx](https://github.com/verena-ifx))
+
+#### 🔩 Dependency Updates
+
+- Bump vue from 3.3.8 to 3.3.10 in /examples/wrapper-components/vue-javascript [#786](https://github.com/Infineon/infineon-design-system-stencil/pull/786) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump @infineon/infineon-design-system-vue from 20.34.5 to 20.36.0 in /examples/wrapper-components/vue-javascript [#785](https://github.com/Infineon/infineon-design-system-stencil/pull/785) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+- Bump @infineon/infineon-design-system-react from 20.34.5 to 20.36.0 in /examples/wrapper-components/react-javascript [#782](https://github.com/Infineon/infineon-design-system-stencil/pull/782) ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+#### Authors: 4
+
+- [@dependabot[bot]](https://github.com/dependabot[bot])
+- [@verena-ifx](https://github.com/verena-ifx)
+- Tisho Yanchev ([@tishoyanchev](https://github.com/tishoyanchev))
+- verena-ifx (lechneve@ISCN5CG1201S7M.infineon.com)
+
+---
+
 # v20.36.1 (Tue Dec 05 2023)
 
 #### 🐛 Bug Fix